# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Garden Sensors API
  version: 1.0.0
  description: API for managing garden sensors and their readings
servers:
  - url: http://localhost/api
    description: Local development server
  - url: https://api.gardensensors.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Sensor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
          enum: [moisture, temperature, humidity, light]
        location:
          type: string
        min_threshold:
          type: number
        max_threshold:
          type: number
        unit:
          type: string
        last_reading:
          type: number
        last_reading_time:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
paths:
  /sensors:
    get:
      summary: List all sensors
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of sensors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sensor'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      summary: Create a new sensor
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sensor'
      responses:
        '201':
          description: Sensor created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sensor'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...